allow runtime choice of uint8 or uint16 for distances matrix element type
- template `DataSet` and `from_fasta` on DistIntType
  - move some code into headers to allow this
- set the default type to `uint8` for all other functions that return a `DataSet`
- modify python interface
  - add a `DataSetLarge` which is just a `DataSet` that uses `uint16` for distances
  - add `from_fasta_large` that uses `uint16` for distances and returns a `DataSetLarge`
  - `from_fasta` and all other existing functions / objects use `uint8` for distances
- update tests & readme
- bump version to 0.18.0
