+++ distance.cc
@@ -9,7 +9,8 @@ int main(int argc, char *argv[]) {
   std::size_t nsamples = std::stoi(std::string(argv[2]));
 
   auto start = std::chrono::steady_clock::now();
-  auto data = hamming::from_fasta(filename, nsamples);
+  auto data =
+      hamming::from_fasta<hamming::DefaultDistIntType>(filename, nsamples);
   data.dump("distances.csv");
   auto stop = std::chrono::steady_clock::now();
   std::chrono::duration<double> elapsed_seconds = stop - start;
