+++ include/hamming/hamming_impl_types.hh
