+++ python/hammingdist.cc
@@ -28,26 +28,49 @@ namespace hamming {
 PYBIND11_MODULE(hammingdist, m) {
   m.doc() = "Small tool to calculate Hamming distances between gene sequences";
 
-  py::class_<DataSet>(m, "DataSet")
-      .def("dump", &DataSet::dump, "Dump distances matrix in csv format")
-      .def("dump_lower_triangular", &DataSet::dump_lower_triangular,
+  py::class_<DataSet<DefaultDistIntType>>(m, "DataSet")
+      .def("dump", &DataSet<DefaultDistIntType>::dump,
+           "Dump distances matrix in csv format")
+      .def("dump_lower_triangular",
+           &DataSet<DefaultDistIntType>::dump_lower_triangular,
            "Dump distances matrix in lower triangular format (comma-delimited, "
            "row-major)")
-      .def("dump_sequence_indices", &DataSet::dump_sequence_indices,
+      .def("dump_sequence_indices",
+           &DataSet<DefaultDistIntType>::dump_sequence_indices,
            "Dump row index in distances matrix for each input sequence")
-      .def("__getitem__", &DataSet::operator[])
-      .def_readonly("_distances", &DataSet::result);
+      .def("__getitem__", &DataSet<DefaultDistIntType>::operator[])
+      .def_readonly("_distances", &DataSet<DefaultDistIntType>::result);
+
+  py::class_<DataSet<uint16_t>>(m, "DataSetLarge")
+      .def("dump", &DataSet<uint16_t>::dump,
+           "Dump distances matrix in csv format")
+      .def("dump_lower_triangular", &DataSet<uint16_t>::dump_lower_triangular,
+           "Dump distances matrix in lower triangular format (comma-delimited, "
+           "row-major)")
+      .def("dump_sequence_indices", &DataSet<uint16_t>::dump_sequence_indices,
+           "Dump row index in distances matrix for each input sequence")
+      .def("__getitem__", &DataSet<uint16_t>::operator[])
+      .def_readonly("_distances", &DataSet<uint16_t>::result);
 
   m.def("from_stringlist", &from_stringlist,
         "Creates a dataset from a list of strings");
   m.def("from_csv", &from_csv,
         "Creates a dataset by reading already computed distances from csv "
         "(full matrix expected)");
-  m.def("from_fasta", &from_fasta, py::arg("filename"),
+  m.def("from_fasta", &from_fasta<uint8_t>, py::arg("filename"),
         py::arg("include_x") = false, py::arg("remove_duplicates") = false,
         py::arg("n") = 0,
         "Creates a dataset by reading from a fasta file (assuming all "
-        "sequences have equal length)");
+        "sequences have equal length). Maximum value of an element in the "
+        "distances matrix: 255. Distances that would have been larger than "
+        "this value saturate at 255 - to support genomes with larger distances "
+        "than this see `from_fasta_large` instead.");
+  m.def("from_fasta_large", &from_fasta<uint16_t>, py::arg("filename"),
+        py::arg("include_x") = false, py::arg("remove_duplicates") = false,
+        py::arg("n") = 0,
+        "Creates a dataset by reading from a fasta file (assuming all "
+        "sequences have equal length). Maximum value of an element in the "
+        "distances matrix: 65535");
   m.def("from_lower_triangular", &from_lower_triangular,
         "Creates a dataset by reading already computed distances from lower "
         "triangular format");
