+++ src/CMakeLists.txt
@@ -1,7 +1,7 @@
 # Build hamming library
 add_library(hamming STATIC hamming.cc hamming_impl.cc)
 target_include_directories(hamming PUBLIC ../include)
-target_link_libraries(hamming PRIVATE CpuFeature::cpu_features)
+target_link_libraries(hamming PUBLIC CpuFeature::cpu_features)
 if(HAMMING_WITH_OPENMP)
   find_package(OpenMP REQUIRED)
   target_compile_definitions(hamming PUBLIC HAMMING_WITH_OPENMP)
@@ -14,7 +14,7 @@ if(HAMMING_WITH_SSE2)
   target_compile_definitions(hamming PUBLIC HAMMING_WITH_SSE2)
   add_library(distance_sse2 STATIC distance_sse2.cc)
   target_compile_options(distance_sse2 PRIVATE -msse2)
-  target_include_directories(distance_sse2 PUBLIC .)
+  target_include_directories(distance_sse2 PUBLIC ../include)
   target_link_libraries(hamming PRIVATE distance_sse2)
 endif()
 
@@ -22,7 +22,7 @@ if(HAMMING_WITH_AVX2)
   target_compile_definitions(hamming PUBLIC HAMMING_WITH_AVX2)
   add_library(distance_avx2 STATIC distance_avx2.cc)
   target_compile_options(distance_avx2 PRIVATE -mavx2)
-  target_include_directories(distance_avx2 PUBLIC .)
+  target_include_directories(distance_avx2 PUBLIC ../include)
   target_link_libraries(hamming PRIVATE distance_avx2)
 endif()
 
@@ -30,7 +30,7 @@ if(HAMMING_WITH_AVX512)
   target_compile_definitions(hamming PUBLIC HAMMING_WITH_AVX512)
   add_library(distance_avx512 STATIC distance_avx512.cc)
   target_compile_options(distance_avx512 PRIVATE -mavx512bw)
-  target_include_directories(distance_avx512 PUBLIC .)
+  target_include_directories(distance_avx512 PUBLIC ../include)
   target_link_libraries(hamming PRIVATE distance_avx512)
 endif()
