+++ src/distance_sse2_t.cc
@@ -1,4 +1,4 @@
-#include "distance_sse2.hh"
+#include "hamming/distance_sse2.hh"
 #include "tests.hh"
 
 using namespace hamming;
