+++ src/hamming_impl_bench.cc
@@ -1,6 +1,6 @@
 #include "bench.hh"
 #include "hamming/hamming.hh"
-#include "hamming_impl.hh"
+#include "hamming/hamming_impl.hh"
 #ifdef HAMMING_WITH_OPENMP
 #include <omp.h>
 #endif
